% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_utils.R
\name{center_from_grid_code}
\alias{center_from_grid_code}
\title{Returns details about the center of a regular grid provided the grid code.
Details include:
    the center latitude / longitude (regardless of the fraction of ocean area in the grid)
    the grid width and height (in degrees)}
\usage{
center_from_grid_code(grid_code)
}
\arguments{
\item{grid_code}{A grid code}
}
\value{
A named vector containing the center latitude (\code{y}) and longitude (\code{x}) plus the grid
width (\code{size_lat}) and height (\code{size_lon}) in degrees
}
\description{
Returns details about the center of a regular grid provided the grid code.
Details include:
    the center latitude / longitude (regardless of the fraction of ocean area in the grid)
    the grid width and height (in degrees)
}
\examples{
center_from_grid_code("5206066")
center_from_grid_code("6205065")
}
