% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_utils.R
\name{convert_to_CWP_grid}
\alias{convert_to_CWP_grid}
\title{Converts a pair of decimal coordinate into a CWP grid code for a specific grid type}
\usage{
convert_to_CWP_grid(lon, lat, grid_type_code = grid_5x5)
}
\arguments{
\item{lon}{Longitude (decimal coordinates)}

\item{lat}{Latitude (decimal coordinates)}

\item{grid_type_code}{The type of CWP grid (one among \code{\link{grid_1x1}}, \code{\link{grid_5x5}}, \code{\link{grid_10x10}}, \code{\link{grid_10x20}}, \code{\link{grid_20x20}} and \code{\link{grid_30x30}})}
}
\value{
The CWP grid code for the provided coordinates and grid type
}
\description{
Converts a pair of decimal coordinate into a CWP grid code for a specific grid type
}
\examples{
convert_to_CWP_grid(20, -10, grid_1x1)
}
