% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_utils.R
\name{CWP_to_grid_coordinates}
\alias{CWP_to_grid_coordinates}
\title{Converts a CWP grid code into its four boundary points (NW, NE, SW and SE)}
\usage{
CWP_to_grid_coordinates(grid_code)
}
\arguments{
\item{grid_code}{A CWP grid code}
}
\value{
a data table containing the coordinates (LAT, LON) of each of the four boundary points for the grid
}
\description{
Converts a CWP grid code into its four boundary points (NW, NE, SW and SE)
}
\examples{
CWP_to_grid_coordinates("5201123")
CWP_to_grid_coordinates("6205125")
}
