% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_10x20}
\alias{grid_10x20}
\title{A constant representing the code for a 10x20 regular grid}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
grid_10x20
}
\description{
A constant representing the code for a 10x20 regular grid
}
\seealso{
grid_code_10x20
}
\keyword{datasets}
