% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_30x30}
\alias{grid_30x30}
\title{A constant representing the code for a 30x30 regular grid}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
grid_30x30
}
\description{
A constant representing the code for a 30x30 regular grid
}
\seealso{
grid_code_30x30
}
\keyword{datasets}
