% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_5x5}
\alias{grid_5x5}
\title{A constant representing the code for a 5x5 regular grid}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
grid_5x5
}
\description{
A constant representing the code for a 5x5 regular grid
}
\seealso{
grid_code_5x5
}
\keyword{datasets}
