% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_ALL}
\alias{grid_ALL}
\title{The list of all constants representing the codes for the various grids}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
grid_ALL
}
\description{
The list of all constants representing the codes for the various grids
}
\keyword{datasets}
