% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_char_1x1}
\alias{grid_char_1x1}
\title{A constant representing the initial character of a 1x1 regular grid code}
\format{
An object of class \code{character} of length 1.
}
\usage{
grid_char_1x1
}
\description{
A constant representing the initial character of a 1x1 regular grid code
}
\keyword{datasets}
