% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_char_20x20}
\alias{grid_char_20x20}
\title{A constant representing the initial character of a 20x20 regular grid code}
\format{
An object of class \code{character} of length 1.
}
\usage{
grid_char_20x20
}
\description{
A constant representing the initial character of a 20x20 regular grid code
}
\keyword{datasets}
