% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_code_1x1}
\alias{grid_code_1x1}
\title{A constant representing the code for a 1x1 regular grid}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
grid_code_1x1
}
\description{
A constant representing the code for a 1x1 regular grid
}
\seealso{
grid_1x1
}
\keyword{datasets}
