% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_cwp_constants.R
\docType{data}
\name{grid_code_irregular}
\alias{grid_code_irregular}
\title{A constant representing the code for an irregular grid}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
grid_code_irregular
}
\description{
A constant representing the code for an irregular grid
}
\seealso{
grid_irregular
}
\keyword{datasets}
